简化了移植过程，现在只需要把global和utility中的#define LINUX_OS注释掉就可以比较容易的移植到mdk上，并且已经移植成功

+++ libraries/rc_channel.cpp
@@ -104,7 +104,7 @@ AP_RC_Channel::set_pwm(int pwm)
 int
 AP_RC_Channel::control_mix(float value)
 {
-	return (1 - fabs(control_in / _high)) * value + control_in;
+	return (1 - fabs((float)control_in /(float) _high)) * value + control_in;
 }
 
 // are we below a threshold?
