+++ location.cpp
@@ -55,7 +55,7 @@ int32_t get_bearing_cd(const struct Location *loc1, const struct Location *loc2)
 {
     int32_t off_x = loc2->lng - loc1->lng;
     int32_t off_y = (loc2->lat - loc1->lat) / longitude_scale(loc2);
-    int32_t bearing = 9000 + atan2(-off_y, off_x) * 5729.57795;
+    int32_t bearing = 9000 + atan2((double)-off_y, (double)off_x) * 5729.57795;
     if (bearing < 0) bearing += 36000;
     return bearing;
 }
